LayoutTests/http/tests/resourceLoadStatistics/clear-in-memory-and-persistent-store-expected.txt

Tests clearing of website data store.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Host classified as prevalent resource.
PASS Host is no longer a prevalent resource after the store was cleared.
PASS successfullyParsed is true

TEST COMPLETE
